Smarter reporting on story table errors
The in-test prints are already a lifesaver for me: it prints the fields
on a page that haven't been set, along with their values. It helps narrow down
which story table might be causing problems, specificly with mispellings and
such. Likely, we'll have to improve by adding it to the report somehow, instead
of printing mid test (though that's also very helpful to me when iterating
locally quickly).

The other change is less well formed, but was trying to track when a variable in
the story table matches something on the page, but the value doesn't.